fix: split context files to fix React Fast Refresh violations causing HMR redirect loops
Context files (Auth, Sdk, Telemetry, CommandPalette, insights-sidebar) exported
both React components and hooks, which violates React Fast Refresh rules. During
HMR, this caused full module re-evaluation instead of component-level hot updates,
briefly losing auth state and triggering redirect loops through the mock IDP.

Split each into a hooks-only file (original path, zero import changes for 130+
consumers) and a component-only file (new *Provider.tsx, only App.tsx updated).
