+++ client/dashboard/src/contexts/Telemetry.tsx
@@ -1,18 +1,13 @@
-import { getServerURL } from "@/lib/utils";
-import posthog, { PostHog } from "posthog-js";
-import { createContext, ReactNode, useContext, useEffect } from "react";
-import { User } from "./Auth";
-import { datadogRum } from "@datadog/browser-rum";
-
-// Set this to true to test telemetry locally
-const AM_TESTING_TELEMETRY = false;
+import type { PostHog } from "posthog-js";
+import { createContext, useContext, useEffect } from "react";
+import type { User } from "./Auth";
 
 export type Telemetry = Pick<
   PostHog,
   "isFeatureEnabled" | "capture" | "identify" | "register" | "reset" | "group"
 >;
 
-const nullTelemetry: Telemetry = {
+export const nullTelemetry: Telemetry = {
   isFeatureEnabled: () => false,
   capture: () => ({ uuid: "", event: "", properties: {} }),
   identify: () => {},
@@ -21,12 +16,12 @@ const nullTelemetry: Telemetry = {
   group: () => {},
 };
 
-const devTelemetry: Telemetry = {
+export const devTelemetry: Telemetry = {
   ...nullTelemetry,
   isFeatureEnabled: () => true,
 };
 
-const testTelemetry: Telemetry = {
+export const testTelemetry: Telemetry = {
   capture: (event: string, properties: Record<string, unknown>) => {
     console.log("POSTHOG CAPTURE", event, properties);
     return { uuid: "", event, properties };
@@ -58,53 +53,6 @@ export const TelemetryContext = createContext<Telemetry>(
 );
 
 export const useTelemetry = () => useContext(TelemetryContext);
-
-export const TelemetryProvider = (props: { children: ReactNode }) => {
-  const ph = posthog.init(
-    "phc_hiYSF5Axu49I1xs4Z5BG8KCI3PGNLM8ERRs7eocmfX9",
-    {
-      api_host: "https://metrics.speakeasy.com",
-      feature_flag_request_timeout_ms: 1000,
-    },
-    "speakeasy",
-  );
-
-  useEffect(() => {
-    // Guard against duplicate initialization (can happen in React StrictMode or component remounts)
-    if (datadogRum.getInitConfiguration()) {
-      return;
-    }
-
-    const serverURL = getServerURL();
-    if (serverURL.includes("getgram.ai")) {
-      const env = serverURL.includes("app.getgram.ai") ? "prod" : "dev";
-      datadogRum.init({
-        applicationId: "93afb64a-dd15-490c-a749-51b4c5c5a171",
-        clientToken: "pub8358667232c624e2f91e1eaa0bd380fd",
-        site: "datadoghq.com",
-        service: "gram",
-        env,
-        sessionSampleRate: 100,
-        sessionReplaySampleRate: 100,
-        trackUserInteractions: true,
-        trackResources: true,
-        defaultPrivacyLevel: "mask-user-input",
-        version: __GRAM_GIT_SHA__ || null,
-      });
-    }
-  }, []);
-
-  let value: Telemetry = ph ?? nullTelemetry;
-  if (getServerURL().includes("localhost")) {
-    value = AM_TESTING_TELEMETRY ? testTelemetry : devTelemetry;
-  }
-
-  return (
-    <TelemetryContext.Provider value={value}>
-      {props.children}
-    </TelemetryContext.Provider>
-  );
-};
 
 export function useIdentifyUserForTelemetry(user: User | undefined) {
   const telemetry = useTelemetry();
